requirements.txt
twitchAPI==4.2.0
python-telegram-bot==21.5
aiohttp==3.10.5
pytz==2024.2